+++ frontend/src/components/@composition/RouletteViewToggle/RouletteViewToggle.tsx
@@ -1,20 +1,18 @@
+import TextButton from '@/components/@common/TextButton/TextButton';
 import { RouletteView } from '@/types/roulette';
-import RouletteIcon from '@/assets/roulette-icon.svg';
-import StatisticsIcon from '@/assets/statistics-icon.svg';
-import IconButton from '@/components/@common/IconButton/IconButton';
 
 type Props = {
   currentView: RouletteView;
   onViewChange: () => void;
 };
 
-const TOGGLE_ICONS = {
-  roulette: StatisticsIcon,
-  statistics: RouletteIcon,
+const TOGGLE_TEXT_MAP = {
+  roulette: '확률 보기',
+  statistics: '룰렛 보기',
 } as const;
 
 const RouletteViewToggle = ({ currentView, onViewChange }: Props) => {
-  return <IconButton iconSrc={TOGGLE_ICONS[currentView]} onClick={onViewChange} />;
+  return <TextButton text={TOGGLE_TEXT_MAP[currentView]} onClick={onViewChange} />;
 };
 
 export default RouletteViewToggle;
