+++ frontend/src/features/room/order/pages/OrderPage.tsx
@@ -1,11 +1,10 @@
 import { useWebSocket } from '@/apis/websocket/contexts/WebSocketContext';
 import BreadLogoWhiteIcon from '@/assets/logo/bread-logo-white.png';
-import DetailIcon from '@/assets/detail-icon.svg';
 import Button from '@/components/@common/Button/Button';
 import Headline1 from '@/components/@common/Headline1/Headline1';
 import Headline2 from '@/components/@common/Headline2/Headline2';
 import Headline3 from '@/components/@common/Headline3/Headline3';
-import IconButton from '@/components/@common/IconButton/IconButton';
+import TextButton from '@/components/@common/TextButton/TextButton';
 import Layout from '@/layouts/Layout';
 import { useEffect, useState } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
@@ -45,7 +44,10 @@ const OrderPage = () => {
       <Layout.Content>
         <S.ListHeader>
           <Headline2>주문 리스트 {viewMode === 'detail' ? '상세' : ''}</Headline2>
-          <IconButton iconSrc={DetailIcon} onClick={handleToggle} />
+          <TextButton
+            text={viewMode === 'detail' ? '요약 보기' : '상세 보기'}
+            onClick={handleToggle}
+          />
         </S.ListHeader>
         {viewMode === 'simple' ? <MenuCount /> : <PlayerMenu />}
       </Layout.Content>
